pyproject.toml
[project]
name = "shlokas-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==3.2.2",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.88",
    "email-validator>=2.3.0",
    "fastapi>=0.128.8",
    "fastapi-cache2[redis]>=0.2.2",
    "google-auth>=2.49.2",
    "httptools>=0.7.1",
    "numpy>=2.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "redis>=4.6.0",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.39.0",
    "uvloop>=0.22.1",
]
